+++ src/styling.ts
@@ -1,23 +1,7 @@
 import ansiRegex from 'ansi-regex';
-import {type ChalkInstance, type ColorSupportLevel} from 'chalk';
+import {type ChalkInstance} from 'chalk';
 
 export * from './browser/styling.js';
-
-/**
- * Contains all color level names.
- */
-export const colorTypeList = [0, 1, 2, 3] as const;
-/**
- * Contains all color level names as a type.
- */
-export type ColorType = ColorSupportLevel;
-/**
- * Checks if the value is the {@link ColorType}.
- */
-export function isColorType(value: unknown): value is ColorType {
-	const numberValue = Number(value);
-	return Number.isFinite(numberValue) && colorTypeList.includes(numberValue as ColorType);
-}
 
 export function highlight(text: string, chalk?: ChalkInstance): string {
 	if (chalk === undefined) {
@@ -63,7 +47,7 @@ export function highlight(text: string, chalk?: ChalkInstance): string {
 		}
 
 		if (match.match(rspecial) !== null) {
-			return chalk.hex('#73DEA7')(match);
+			return chalk.hex('#73A7DE')(match);
 		}
 
 		return match;